separando responsabilidades
