+++ app/graphql/resolvers.js
@@ -1,0 +1,20 @@
+const { getProduct, getProducts } = require("./product.resolvers");
+
+const resolvers = {
+  Query: {
+    hello: () => "Hola mundillo",
+    getPerson: (_, { name, age }) =>
+      `Hello, my name is ${name}, I'm ${age} years old`,
+    getInt: (_, { age }) => age,
+    getFloat: (_, { price }) => price,
+    getString: () => "palabra",
+    getBoolean: () => true,
+    getID: () => "1212",
+    getNumbers: (_, { numbers }) => numbers,
+    // Products
+    product: getProduct,
+    products: getProducts,
+  },
+};
+
+module.exports = resolvers;
