+++ app/graphql/schema.graphql
@@ -7,7 +7,8 @@ type Query {
   getBoolean: Boolean
   getID: ID
   getNumbers(numbers: [Int!]!): [Int]
-  getProduct: Product!
+  product(id: ID!): Product
+  products: [Product!]!
 }
 
 type Product {
